Medaccess CoPD import + make isNumeric check accept floating point numbers.

+++ src/main/java/org/oscarehr/demographicImport/service/CoPDPreProcessorService.java
@@ -255,7 +255,7 @@ public class CoPDPreProcessorService
 			{
 				try
 				{
-					Integer.parseInt(tagValue);
+					Float.parseFloat(tagValue);
 					return tagValue;
 				}
 				catch (NumberFormatException e)
